backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
python-decouple>=3.8
psycopg2-binary>=2.9.9
djangorestframework-simplejwt>=5.3.1
openpyxl>=3.1.2
Pillow>=10.2.0
django-axes>=6.2.0
requests>=2.31.0
